Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
sqlmodel==0.0.22
pydantic==2.10.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1